split up modals, add placeholder progress modal
