+++ src/scenes/board/board-controls.tsx
@@ -48,6 +48,25 @@ const StyledHelpButton = styled(StyledButton)`
   }
 `;
 
+const StyledProgressButton = styled(StyledButton)`
+  padding: 0.25rem .5rem;
+  color: ${getColor('purple')};
+  background-color:${getColor('brown_dark')};
+  border-color: ${getColor('purple')};
+  box-shadow: 0.4rem 0.4rem 0 0.1rem ${getColor('purple')};
+  
+  &:hover{
+    color: ${getColor('pink')};
+    border-color: ${getColor('pink')};
+  }
+  &:active{
+    color: ${getColor('purple')};
+    background-color: ${getColor('brown_dark')};
+    border-color: ${getColor('purple')};
+    box-shadow: 0.1rem 0.1rem 0 0.3rem ${getColor('purple')};
+  }
+`;
+
 const StyledContainer = styled.div`
   ${StyledButton}{
     margin:1rem;
@@ -64,6 +83,7 @@ export function BoardControls() {
 
   return (
     <StyledContainer>
+      <StyledProgressButton onClick={() => dispatch(setGameStatus('progress'))}>{'progress'}</StyledProgressButton>
       <StyledHelpButton onClick={() => dispatch(setGameStatus('debug'))}>{'debug'}</StyledHelpButton>
       <StyledHelpButton onClick={() => dispatch(setGameStatus('help'))}>{'HELP!'}</StyledHelpButton>
       <StyledSolvedButton onClick={() => onSubmitGame(solved)}>{'SUBMIT'}</StyledSolvedButton>
